ml_service/requirements.txt
tensorflow==2.15.0
flask
opencv-python
numpy
requests
scikit-learn
pandas
pillow